# agent-api/openapi.yaml
openapi: 3.0.0
info:
  title: Task Management API
  description: A simple task management system for demonstrating Willknow Agent capabilities
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/tasks:
    get:
      operationId: listTasks
      summary: List all tasks
      description: Returns a list of all tasks in the system
      parameters:
        - name: status
          in: query
          description: Filter tasks by status
          required: false
          schema:
            type: string
            enum: [pending, in_progress, completed]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'

    post:
      operationId: createTask
      summary: Create a new task
      description: Creates a new task with the provided information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: Task title
                description:
                  type: string
                  description: Detailed task description
                priority:
                  type: string
                  enum: [low, medium, high]
                  description: Task priority
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /api/tasks/{taskId}:
    get:
      operationId: getTask
      summary: Get a specific task
      description: Returns details of a single task by ID
      parameters:
        - name: taskId
          in: path
          required: true
          description: The task ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

    put:
      operationId: updateTask
      summary: Update a task
      description: Updates an existing task's information
      parameters:
        - name: taskId
          in: path
          required: true
          description: The task ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Task title
                description:
                  type: string
                  description: Detailed task description
                status:
                  type: string
                  enum: [pending, in_progress, completed]
                  description: Task status
                priority:
                  type: string
                  enum: [low, medium, high]
                  description: Task priority
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

    delete:
      operationId: deleteTask
      summary: Delete a task
      description: Removes a task from the system
      parameters:
        - name: taskId
          in: path
          required: true
          description: The task ID
          schema:
            type: string
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found

  /api/tasks/{taskId}/complete:
    post:
      operationId: completeTask
      summary: Mark task as completed
      description: Marks a task as completed and records completion time
      parameters:
        - name: taskId
          in: path
          required: true
          description: The task ID
          schema:
            type: string
      responses:
        '200':
          description: Task marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          description: Unique task identifier
        title:
          type: string
          description: Task title
        description:
          type: string
          description: Detailed task description
        status:
          type: string
          enum: [pending, in_progress, completed]
          description: Current task status
        priority:
          type: string
          enum: [low, medium, high]
          description: Task priority
        createdAt:
          type: string
          format: date-time
          description: Task creation timestamp
        completedAt:
          type: string
          format: date-time
          description: Task completion timestamp (null if not completed)